Update eventHandler.cfc
We've run into conflicts because the external cache is being used for every site, not just the sites that have this plugin assigned. This has caused some issues and confusion. Let's discuss further is this might have some wider implications but I'm in the process of testing this change right now.